- Organized project. - Refactored some items.
