+++ MyApplication/.idea/.name
@@ -1,1 +1,1 @@
-BLPAppv2
+MyApplication
